Move "table qualified with the session catalog name"
